+++ nengo/builder/connection.py
@@ -295,7 +295,7 @@ def build_connection(model, conn):
 
     # Build synapse
     if conn.synapse is not None:
-        weighted = model.build(conn.synapse, weighted, mode="update")
+        weighted = model.build(conn.synapse, weighted, mode="update", seed_or_rng=rng)
 
     # Store the weighted-filtered output in case we want to probe it
     model.sig[conn]["weighted"] = weighted